fixed code

+++ func.h
@@ -1,5 +1,5 @@
 #pragma once
 
-double geometricAverage(double a, double b);
-double func1(int x);
+double geometricAverage(int a, double b);
+int func1(int x);
 double func2(double x, double y);
